+++ topobank_publication/views.py
@@ -20,7 +20,6 @@ from .serializers import PublicationSerializer
 from .utils import NewPublicationTooFastException, PublicationException
 
 from topobank.manager.models import Surface
-from topobank.manager.views import download_surface
 from topobank.usage_stats.utils import increase_statistics_by_date_and_object
 
 _log = logging.getLogger(__name__)
@@ -65,17 +64,11 @@ def go(request, short_url):
         raise Http404()
 
     increase_statistics_by_date_and_object(Metric.objects.PUBLICATION_VIEW_COUNT, period=Period.DAY, obj=pub)
-    return redirect(pub.surface.get_absolute_url())
 
-
-def download(request, short_url):
-    """Download a published surface by short url."""
-    try:
-        pub = Publication.objects.get(short_url=short_url)
-    except Publication.DoesNotExist:
-        raise Http404()
-
-    return download_surface(request, pub.surface_id)
+    if 'HTTP_ACCEPT' in request.META and 'application/json' in request.META['HTTP_ACCEPT']:
+        return redirect(pub.get_api_url())
+    else:
+        return redirect(f"{reverse('ce_ui:surface-detail')}?surface={pub.surface.pk}")  # <- topobank does not know this
 
 
 class SurfacePublishView(FormView):