+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog for plugin *topobank-publication*
 
-## 1.5.1 (not yet released)
+## 1.6.0 (2024-01-26)
 
+- ENH: /go links return API redirect if `application/json` is requested,
+  otherwise HTML redirect (#9)
+- ENH: API endpoint for publication now returns download link
+- BUG: Fix to /go links (#8)
 - MAINT: Adding gitignore
 
 ## 1.5.0 (2024-01-20)